requirements.txt
streamlit
pandas
numpy
scipy
statsmodels
sqlalchemy
psycopg2-binary
openpyxl
